# tools/swagger/docs/api-internal/iam-internal/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ces APIS sont utilisés pour configurer les clients, tenants, idp, groupes, profils, et utilisateurs.",
    "version": "v1",
    "title": "API IAM - Gestion des accès aux identités",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Direction de la diffusion et des partenariats",
      "email": "contact@programmevitam.fr"
    },
    "license": {
      "name": "License to be defined ...",
      "url": "http://notedefined.com"
    }
  },
  "host": "localhost",
  "basePath": "/iam/v1",
...
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/customers/{id}/history": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Récupération de l'historique d'un client",
        "operationId": "findHistoryByIdUsingGET",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du client",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/groups/{id}/history": {
      "get": {
        "tags": [
          "Groupes"
        ],
        "summary": "Récupération de l'historique d'un groupe",
        "operationId": "findHistoryByIdUsingGET_1",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du groupe",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/owners/{id}/history": {
      "get": {
        "tags": [
          "Propriétaires"
        ],
        "summary": "Récupération de l'historique d'un propriétaire",
        "operationId": "findHistoryByIdUsingGET_2",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du propriétaire",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/profiles/{id}/history": {
      "get": {
        "tags": [
          "Profils"
        ],
        "summary": "Récupération de l'historique d'un profil",
        "operationId": "findHistoryByIdUsingGET_3",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique du profil",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
...
    "/users/{id}/history": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération de l'historique d'un utilisateur",
        "operationId": "findHistoryByIdUsingGET_4",
        "parameters": [
          {
            "$ref": "#/parameters/x-tenant-id"
          },
          {
            "$ref": "#/parameters/x-user-token"
          },
          {
            "$ref": "#/parameters/x-application-id"
          },
          {
            "$ref": "#/parameters/x-user-level"
          },
          {
            "$ref": "#/parameters/x-customer-id"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant technique de l'utilisateur",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Historique d'un utilisateur",
            "schema": {
              "$ref": "#/definitions/JsonNode"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    }
  },
  "responses": {
    "Unauthorized": {
      "description": "La requête n'est pas autorisée. Le X-User-Token n'est pas valide"
    },
    "NotFound": {
      "description": "Ressource non trouvée"
    }
  },
  "parameters": {
...
    "x-tenant-id": {
      "name": "X-Tenant-Id",
      "in": "header",
      "description": "Identifiant du tenant sur lequel doit être exécuté la requête",
      "required": true,
      "type": "integer",
      "format": "int32"
    },
    "x-user-token": {
      "name": "X-User-Token",
      "in": "header",
      "description": "Identifiant technique du token généré par le serveur CAS lors de l'authentification",
      "required": true,
      "type": "string"
    },
    "x-application-id": {
      "name": "X-Application-Id",
      "in": "header",
      "description": "Identifiant unique utilisé dans la génération des logs et des journaux VITAM",
      "required": false,
      "type": "string"
    },
    "x-user-level": {
      "name": "X-User-Level",
      "in": "header",
      "description": "Niveau de l'utilisateur",
      "required": true,
      "type": "string"
    },
    "x-customer-id": {
      "name": "X-Customer-Id",
      "in": "header",
      "description": "Identifiant technique du client",
      "required": true,
      "type": "string"
    }
  },
  "definitions": {
...
    "JsonNode": {
      "type": "object",
      "properties": {
        "array": {
          "type": "boolean"
        },
        "bigDecimal": {
          "type": "boolean"
        },
        "bigInteger": {
          "type": "boolean"
        },
        "binary": {
          "type": "boolean"
        },
        "boolean": {
          "type": "boolean"
        },
        "containerNode": {
          "type": "boolean"
        },
        "double": {
          "type": "boolean"
        },
        "float": {
          "type": "boolean"
        },
        "floatingPointNumber": {
          "type": "boolean"
        },
        "int": {
          "type": "boolean"
        },
        "integralNumber": {
          "type": "boolean"
        },
        "long": {
          "type": "boolean"
        },
        "missingNode": {
          "type": "boolean"
        },
        "nodeType": {
          "type": "string",
          "enum": [
            "ARRAY",
            "BINARY",
            "BOOLEAN",
            "MISSING",
            "NULL",
            "NUMBER",
            "OBJECT",
            "POJO",
            "STRING"
          ]
        },
        "null": {
          "type": "boolean"
        },
        "number": {
          "type": "boolean"
        },
        "object": {
          "type": "boolean"
        },
        "pojo": {
          "type": "boolean"
        },
        "short": {
          "type": "boolean"
        },
        "textual": {
          "type": "boolean"
        },
        "valueNode": {
          "type": "boolean"
        }
      },
      "title": "JsonNode"
    },
...
  }
}